feat: Добавить контроль прав на удаление карточки

+++ controllers/cards.js
@@ -1,7 +1,8 @@
 const Card = require("../models/card");
-const NotFoundError = require("../utils/errors");
+const {NotFoundError} = require("../utils/errors");
 
 module.exports.getCards = (req, res, next) => {
+  console.log(req.user._id)
   Card.find({})
     .then((cards) => {
       return res.send({ data: cards });
@@ -17,7 +18,8 @@ module.exports.createCard = (req, res, next) => {
 };
 
 module.exports.deleteCard = (req, res, next) => {
-  Card.findOneAndDelete({ _id: req.params.cardId })
+
+  Card.findOneAndDelete({ _id: req.params.cardId, owner:req.user._id })
     .then((card) => {
       if (!card) {
         throw new NotFoundError("Запрашиваемая карта не найдена");
